test(migrate): move D1 migration tests from "cli" to "migrate"
