+++ packages/cli/package.json
@@ -111,7 +111,6 @@
     "strip-ansi": "6.0.1",
     "ts-pattern": "5.1.1",
     "typescript": "5.4.5",
-    "wrangler": "3.56.0",
     "xdg-app-paths": "8.3.0",
     "zx": "7.2.3"
   },
