+++ packages/migrate/package.json
@@ -48,7 +48,8 @@
     "strip-indent": "3.0.0",
     "tempy": "1.0.1",
     "ts-pattern": "5.1.1",
-    "typescript": "5.4.5"
+    "typescript": "5.4.5",
+    "wrangler": "3.57.1"
   },
   "peerDependencies": {
     "@prisma/generator-helper": "*",
