+++ packages/migrate/src/__tests__/fixtures/schema-sqlite-d1-24208/prisma/schema.prisma
