+++ packages/migrate/src/__tests__/fixtures/schema-sqlite-d1-change-all-referenced-columns/prisma/schema-0002_change_all_referenced_columns.prisma
