+++ packages/migrate/src/__tests__/fixtures/schema-sqlite-d1-change-all-referenced-columns/prisma/schema.prisma
