+++ packages/migrate/src/__tests__/fixtures/schema-sqlite-d1-migrations-do-not-cause-data-loss/prisma/schema-0002_passport_number_to_string.prisma
