+++ packages/migrate/src/__tests__/fixtures/schema-sqlite-d1-mutual-references/prisma/schema-0002_add_mutual_references.prisma
