+++ packages/migrate/src/__tests__/fixtures/schema-sqlite-d1-mutual-references/prisma/schema.prisma
