+++ packages/migrate/src/__tests__/fixtures/schema-sqlite-d1/prisma/schema-0002_add_count_to_user_table.prisma
