+++ packages/migrate/src/__tests__/fixtures/schema-sqlite-d1/prisma/schema-0003_change_user_id_to_count.prisma
