+++ packages/migrate/src/__tests__/fixtures/schema-sqlite-d1/prisma/schema.prisma
